d/rules: Enable libsystemd dependency, for logind integration
This appears to have got lost in the switch to the Meson build system,
where both libsystemd and libelogind support for communicating with a
logind implementation are disabled by default (they are mutually exclusive).

In particular, this fixes the "Switch User..." menu option in GNOME
Shell, which disappeared because libaccountsservice didn't tell GNOME
Shell that user-switching was possible.

This should work equally well for elogind due to libelogind's libsystemd
ABI compatibility, although I haven't tried that configuration.

Closes: #968394